requirements.txt
langchain>=0.3
langchain-openai>=0.3
langchain-community>=0.3
python-dotenv>=1.0
fastapi>=0.115
uvicorn>=0.34
faiss-cpu>=1.7
sentence-transformers>=2.2
